#5324 remove any traces from Lombok
