--- processor/src/test/java/org/mapstruct/ap/test/lombok/package-info.java
@@ -1,9 +1,0 @@
-/*
- * Copyright MapStruct Authors.
- *
- * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
- */
-/**
- * This package contains tests with delomboked classes.
- */
-package org.mapstruct.ap.test.lombok;
