+++ processor/src/test/java/org/mapstruct/ap/test/superbuilder/AbstractVehicle.java
@@ -3,19 +3,8 @@
  *
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.ap.test.lombok.superbuilder;
+package org.mapstruct.ap.test.superbuilder;
 
-/**
- * Delomboked version of an abstract super class using a <code>lombok.experimental.SuperBuilder</code>.
- * <p>
- * <pre>
- * &#64;SuperBuilder
- * &#64;Getter
- * public abstract class AbstractVehicle {
- *     private final int amountOfTires;
- * }
- * </pre>
- */
 public abstract class AbstractVehicle {
 
     private final int amountOfTires;
@@ -34,10 +23,9 @@ public abstract class AbstractVehicle {
         return this.passenger;
     }
 
-    // CHECKSTYLE:OFF
-    public static abstract class AbstractVehicleBuilder<C extends AbstractVehicle,
+    public abstract static class AbstractVehicleBuilder<C extends AbstractVehicle,
         B extends AbstractVehicleBuilder<C, B>> {
-        // CHECKSTYLE:ON
+
         private int amountOfTires;
         private Passenger passenger;
 
@@ -54,10 +42,5 @@ public abstract class AbstractVehicle {
         protected abstract B self();
 
         public abstract C build();
-
-        public String toString() {
-            return "AbstractVehicle.AbstractVehicleBuilder(amountOfTires=" + this.amountOfTires + ", passenger=" +
-                this.passenger + ")";
-        }
     }
 }
