+++ processor/src/test/java/org/mapstruct/ap/test/superbuilder/MuscleCarDto.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.ap.test.lombok.superbuilder;
+package org.mapstruct.ap.test.superbuilder;
 
 public class MuscleCarDto extends CarDto {
