+++ processor/src/test/java/org/mapstruct/ap/test/superbuilder/Passenger.java
@@ -3,19 +3,8 @@
  *
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.ap.test.lombok.superbuilder;
+package org.mapstruct.ap.test.superbuilder;
 
-/**
- * Delomboked version of a class without inheritance using a <code>lombok.experimental.SuperBuilder</code>.
- * <p>
- * <pre>
- * &#64;SuperBuilder
- * &#64;Getter
- * public class Passenger {
- *     private final String name;
- * }
- * </pre>
- */
 public class Passenger {
 
     private final String name;
@@ -32,9 +21,8 @@ public class Passenger {
         return this.name;
     }
 
-    // CHECKSTYLE:OFF
-    public static abstract class PassengerBuilder<C extends Passenger, B extends PassengerBuilder<C, B>> {
-        // CHECKSTYLE:ON
+    public abstract static class PassengerBuilder<C extends Passenger, B extends PassengerBuilder<C, B>> {
+
         private String name;
 
         public B name(String name) {
@@ -45,10 +33,6 @@ public class Passenger {
         protected abstract B self();
 
         public abstract C build();
-
-        public String toString() {
-            return "Passenger.PassengerBuilder(name=" + this.name + ")";
-        }
     }
 
     private static final class PassengerBuilderImpl extends PassengerBuilder<Passenger, PassengerBuilderImpl> {
